+++ flake.systems.nix
@@ -1,1 +1,0 @@
-[ "x86_64-linux" "aarch64-darwin" "aarch64-linux" ]