+++ hosts/gaia/home.nix
@@ -1,1 +1,8 @@
-{ }
+{
+  wayland.windowManager.sway.config.workspaceOutputAssign = [{
+    workspace = "1";
+    output = "VGA-1";
+  }];
+
+  wayland.windowManager.sway.config.startup = [{ command = "firefox"; }];
+}