+++ hosts/hyperion/darwin-configuration.nix
@@ -35,7 +35,8 @@
       sshUser = "builder";
       sshKey = config.clan.core.vars.generators.nix-remote-build.files.key.path;
       system = "x86_64-linux";
-      supportedFeatures = [ "kvm" "benchmark" "big-parallel" "nixos-test" ];
+      supportedFeatures =
+        [ "kvm" "benchmark" "big-parallel" "nixos-test" "uid-range" ];
       maxJobs = 96;
     }
     {
@@ -44,7 +45,8 @@
       sshUser = "builder";
       sshKey = config.clan.core.vars.generators.nix-remote-build.files.key.path;
       system = "aarch64-linux";
-      supportedFeatures = [ "kvm" "benchmark" "big-parallel" "nixos-test" ];
+      supportedFeatures =
+        [ "kvm" "benchmark" "big-parallel" "nixos-test" "uid-range" ];
       maxJobs = 96;
     }
   ];