+++ modules/flake-parts/devShells.nix
@@ -3,8 +3,11 @@
     devShells.default = pkgs.mkShell {
       buildInputs = builtins.attrValues {
         inherit (inputs'.home-manager.packages) home-manager;
-        inherit (inputs'.clan-core.packages) clan-cli;
         inherit (self'.packages) tf;
+
+        clan-cli = inputs'.clan-core.packages.clan-cli.override {
+          nix = pkgs.lixPackageSets.latest.lix;
+        };
       };
 
       shellHook = ''