+++ modules/flake-parts/firefox-addons.nix
@@ -1,18 +1,14 @@
+{ inputs, ... }:
+
 {
-  inputs.nixpkgs.url = "github:NixOS/nixpkgs/nixos-unstable";
-
-  inputs.firefox-addons.url =
-    "gitlab:rycee/nur-expressions?dir=pkgs/firefox-addons";
-  inputs.firefox-addons.inputs.nixpkgs.follows = "nixpkgs";
-
-  outputs = { firefox-addons, ... }: {
-    overlay = final: prev:
+  flake = {
+    overlays.firefox-addons = final: prev:
       let
         # We need to import default.nix to use buildFirefoxXpiAddon which doesn't get exported in flake.nix
         # WORKAROUND: In Nix 2.14+, firefox-addons.outPath points to the subdirectory rather than the root
         #             so we need to use sourceInfo.outPath to maintain backwards compatibility
-        addons =
-          import "${firefox-addons.sourceInfo.outPath}/pkgs/firefox-addons" {
+        addons = import
+          "${inputs.firefox-addons.sourceInfo.outPath}/pkgs/firefox-addons" {
             inherit (prev) fetchurl lib stdenv;
           };
       in {